Update Readme to reflect a new best practice
I found that early this year there has been a change in openhab that allows for a Member of trigger of a rule. This makes the process a lot cleaner (no more rule adaptations, just add the item to the correct group). I propose to change the readme to reflect this.